// src/app/booking/types/booking-status-types.ts
export enum bookingStatusType {
  OpenBooking = 'OpenBooking',
  NewBooking = 'NewBooking',
  RejectBookingByGuest = 'RejectBookingByGuest',
  RejectBookingByHost = 'RejectBookingByHost',
  RejectBookingByOtaghak = 'RejectBookingByOtaghak',
  AcceptBookingByHost = 'AcceptBookingByHost',
  ConfirmBooking = 'ConfirmBooking',
  CancelBookingByGuest = 'CancelBookingByGuest',
  CancelBookingByHost = 'CancelBookingByHost',
  CancelBookingByOtaghak = 'CancelBookingByOtaghak',
  FinishedBooking = 'FinishedBooking',
  UnpaidBooking = 'UnpaidBooking'
}